Add --match-server-url option
This flag allows one to specify how strict the open api server matching
should be. None means that it should ignore the host, protocol etc and
only match paths. This is useful if e.g. your tests target localhost but
you specify a proper url in the open api. Full means that the request
has to fully match what is stated in the specification.

The logic is somewhat convoluted and can be improved. However, this
covers the cases that I tested. We can improve as we further understand
how this tool will be used